+++ src/entities/mod.rs
@@ -1,0 +1,4 @@
+mod physics;
+mod animations;
+pub mod blocks;
+pub mod mobs;
