--- src/init_and_physics/mod.rs
@@ -1,2 +1,0 @@
-pub mod setup;
-mod physics;
