--- src/init_and_physics/setup.rs
@@ -1,63 +1,0 @@
-use bevy::{prelude::*, sprite::MaterialMesh2dBundle};
-use bevy::core_pipeline::clear_color::ClearColorConfig;
-use crate::init_and_physics::physics;
-
-pub fn setup(mut commands: Commands, asset_server: Res<AssetServer>,
-         mut texture_atlases: ResMut<Assets<TextureAtlas>>) {
-    // Camera Settings
-    commands.spawn(Camera2dBundle {
-        camera_2d: Camera2d {
-            clear_color: ClearColorConfig::Custom(Color::rgb(0.0, 0.0, 1.0)),
-        },
-        ..Default::default()
-    });
-    // Asset Settings
-    let texture_handle = asset_server.load("./adventurer_sprite_sheet.png");
-    let texture_atlas =
-        TextureAtlas::from_grid(texture_handle, Vec2::new(32.0, 32.0), 7, 1, None, None);
-    let texture_atlas_handle = texture_atlases.add(texture_atlas);
-    // Use only the subset of sprites in the sheet that make up the run animation
-    let animation_indices = AnimationIndices { first: 1, last: 6 };
-    commands.spawn((
-        SpriteSheetBundle {
-            texture_atlas: texture_atlas_handle,
-            sprite: TextureAtlasSprite::new(animation_indices.first),
-            ..default()
-        },
-        animation_indices,
-        AnimationTimer(Timer::from_seconds(0.2, TimerMode::Repeating)),
-    ));
-    // World Gen
-    commands.spawn(physics::spawn_grass_block(0.,-50.,0.));
-}
-
-// Animation for sprites
-// TODO: Move into own file?
-#[derive(Component)]
-pub struct AnimationIndices {
-    first: usize,
-    last: usize,
-}
-
-#[derive(Component, Deref, DerefMut)]
-pub struct AnimationTimer(Timer);
-
-pub fn animate_sprite(
-    time: Res<Time>,
-    mut query: Query<(
-        &AnimationIndices,
-        &mut AnimationTimer,
-        &mut TextureAtlasSprite,
-    )>,
-) {
-    for (indices, mut timer, mut sprite) in &mut query {
-        timer.tick(time.delta());
-        if timer.just_finished() {
-            sprite.index = if sprite.index == indices.last {
-                indices.first
-            } else {
-                sprite.index + 1
-            };
-        }
-    }
-}
