+++ src/main.rs
@@ -1,12 +1,11 @@
-mod init_and_physics;
-
 use bevy::prelude::*;
-use crate::init_and_physics::setup;
+mod startup;
+mod entities;
 
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_startup_system(setup::setup)
-        .add_system(setup::animate_sprite)
+        .add_startup_system(startup::startup)
+        .add_system(startup::animate_sprite)
         .run();
 }
